@@ -2,7 +2,7 @@
 /* eslint-disable react/no-multi-comp */
 import chai, { expect } from 'chai';
 import dirtyChai from 'dirty-chai';
-import { mount } from 'enzyme';
+import { shallow } from 'enzyme';
 import uniqueId from 'lodash.uniqueid';
 import React from 'react';
 import sinon from 'sinon';
@@ -44,7 +44,7 @@ descriptor('PureComponent extension', () => {
 	const callbackRender = sandbox.spy(TestComponent.prototype, 'render');
 
 	beforeEach(() => {
-		component = mount(<TestComponent {...getUniqueProps()} />);
+		component = shallow(<TestComponent {...getUniqueProps()} />);
 	});
 
 	afterEach(() => sandbox.reset());
@@ -55,7 +55,7 @@ descriptor('PureComponent extension', () => {
 		});
 
 		it('runs on mount with first parameter of component props', () => {
-			expect(callbackWill.firstCall).to.have.been.calledWith(component.props());
+			expect(callbackWill.firstCall).to.have.been.calledWith(component.instance().props);
 		});
 
 		it('runs on mount with "this" context of component', () => {
@@ -72,13 +72,13 @@ descriptor('PureComponent extension', () => {
 		});
 
 		it('runs on props update when no props change', () => {
-			component.setProps(component.props());
+			component.setProps(component.instance().props);
 			expect(callbackWill).to.have.been.calledTwice();
 		});
 
 		it('runs on props update with first parameter of component props', () => {
 			component.setProps(getUniqueProps());
-			expect(callbackWill.secondCall).to.have.been.calledWith(component.props());
+			expect(callbackWill.secondCall).to.have.been.calledWith(component.instance().props);
 		});
 
 		it('runs on props update with "this" context of component', () => {
@@ -103,7 +103,7 @@ descriptor('PureComponent extension', () => {
 		});
 
 		it('runs on mount with first parameter of component props', () => {
-			expect(callbackDid.firstCall).to.have.been.calledWith(component.props());
+			expect(callbackDid.firstCall).to.have.been.calledWith(component.instance().props);
 		});
 
 		it('runs on mount with second parameter of component state', () => {
@@ -124,7 +124,7 @@ descriptor('PureComponent extension', () => {
 		});
 
 		it('runs on props update with first parameter of previous component props', () => {
-			const initialProps = component.props();
+			const initialProps = component.instance().props;
 			component.setProps(getUniqueProps());
 			expect(callbackDid.secondCall).to.have.been.calledWith(initialProps);
 		});
@@ -146,7 +146,7 @@ descriptor('PureComponent extension', () => {
 		});
 
 		it('runs on state update with first parameter of previous component props', () => {
-			const initialProps = component.props();
+			const initialProps = component.instance().props;
 			component.setState(getUniqueState());
 			expect(callbackDid.secondCall).to.have.been.calledWith(initialProps);
 		});
@@ -215,7 +215,7 @@ descriptor('PureComponent extension with overrides calling super()', () => {
 	const callbackDid = sandbox.spy(TestComponentWithSuper.prototype, 'componentDidMountOrUpdate');
 
 	beforeEach(() => {
-		component = mount(<TestComponentWithSuper {...getUniqueProps()} />);
+		component = shallow(<TestComponentWithSuper {...getUniqueProps()} />);
 	});
 
 	afterEach(() => sandbox.reset());
@@ -300,10 +300,13 @@ descriptor('Component extension with overrides not calling super()', () => {
 	const { PureComponent } = require('./'); // eslint-disable-line global-require
 
 	class TestComponentWithoutSuper extends PureComponent {
-		componentWillMount() {}
-		componentDidMount() {}
-		componentWillReceiveProps() {}
-		componentDidUpdate() {}
+		componentWillMount() { }
+
+		componentDidMount() { }
+
+		componentWillReceiveProps() { }
+
+		componentDidUpdate() { }
 
 		render() {
 			return null;
@@ -314,7 +317,7 @@ descriptor('Component extension with overrides not calling super()', () => {
 	const callbackDid = sandbox.spy(TestComponentWithoutSuper.prototype, 'componentDidMountOrUpdate');
 
 	beforeEach(() => {
-		component = mount(<TestComponentWithoutSuper {...getUniqueProps()} />);
+		component = shallow(<TestComponentWithoutSuper {...getUniqueProps()} />);
 	});
 
 	afterEach(() => sandbox.reset());
