@@ -2,7 +2,7 @@
 import chai, { expect } from 'chai';
 import createReactClass from 'create-react-class';
 import dirtyChai from 'dirty-chai';
-import { mount } from 'enzyme';
+import { shallow } from 'enzyme';
 import uniqueId from 'lodash.uniqueid';
 import React from 'react';
 import sinon from 'sinon';
@@ -37,7 +37,7 @@ describe('withEvents extension', () => {
 	const callbackRender = sandbox.spy(TestComponent.prototype, 'render');
 
 	beforeEach(() => {
-		component = mount(<TestComponent {...getUniqueProps()} />);
+		component = shallow(<TestComponent {...getUniqueProps()} />);
 	});
 
 	afterEach(() => sandbox.reset());
@@ -48,7 +48,7 @@ describe('withEvents extension', () => {
 		});
 
 		it('runs on mount with first parameter of component props', () => {
-			expect(callbackWill.firstCall).to.have.been.calledWith(component.props());
+			expect(callbackWill.firstCall).to.have.been.calledWith(component.instance().props);
 		});
 
 		it('runs on mount with "this" context of component', () => {
@@ -65,13 +65,13 @@ describe('withEvents extension', () => {
 		});
 
 		it('runs on props update when no props change', () => {
-			component.setProps(component.props());
+			component.setProps(component.instance().props);
 			expect(callbackWill).to.have.been.calledTwice();
 		});
 
 		it('runs on props update with first parameter of component props', () => {
 			component.setProps(getUniqueProps());
-			expect(callbackWill.secondCall).to.have.been.calledWith(component.props());
+			expect(callbackWill.secondCall).to.have.been.calledWith(component.instance().props);
 		});
 
 		it('runs on props update with "this" context of component', () => {
@@ -96,7 +96,7 @@ describe('withEvents extension', () => {
 		});
 
 		it('runs on mount with first parameter of component props', () => {
-			expect(callbackDid.firstCall).to.have.been.calledWith(component.props());
+			expect(callbackDid.firstCall).to.have.been.calledWith(component.instance().props);
 		});
 
 		it('runs on mount with second parameter of component state', () => {
@@ -117,7 +117,7 @@ describe('withEvents extension', () => {
 		});
 
 		it('runs on props update with first parameter of previous component props', () => {
-			const initialProps = component.props();
+			const initialProps = component.instance().props;
 			component.setProps(getUniqueProps());
 			expect(callbackDid.secondCall).to.have.been.calledWith(initialProps);
 		});
@@ -139,7 +139,7 @@ describe('withEvents extension', () => {
 		});
 
 		it('runs on state update with first parameter of previous component props', () => {
-			const initialProps = component.props();
+			const initialProps = component.instance().props;
 			component.setState(getUniqueState());
 			expect(callbackDid.secondCall).to.have.been.calledWith(initialProps);
 		});
@@ -183,7 +183,7 @@ describe('withEvents extension with overrides', () => {
 	}));
 
 	beforeEach(() => {
-		component = mount(<TestComponentWithOverrides {...getUniqueProps()} />);
+		component = shallow(<TestComponentWithOverrides {...getUniqueProps()} />);
 	});
 
 	afterEach(() => sandbox.reset());
